refactor: 단위 컴포넌트 폴더로 분리
Rename: Content -> EventMonitor
단위 컴포넌트들 폴더를 생성하여 각각의 폴더로 이동
