+++ frontend/host-app/src/components/EventMonitor/Column.js
@@ -1,9 +1,9 @@
 import React, {useState} from "react";
-import Title from "./Title";
-import QuestionContainer from "./Questions/QuestionContainer";
+import Title from "./Title/Title";
+import QuestionContainer from "../Questions/QuestionContainer";
 import ColumnFooter from "./ColumnFooter";
 import {QuestionStyle, ModerationStyle} from "./ComponentsStyle";
-import {filterStared} from "../libs/utils";
+import {filterStared} from "../../libs/utils";
 
 function Column({type, state, stateHandler, data, dataHandler, handleStar}) {
 	const [heightWeight, setHeightWeight] = useState(0);
