+++ frontend/host-app/src/components/EventMonitor/ComponentsStyle.js
