+++ frontend/host-app/src/components/Header/HeaderConfigAvata.js
