+++ frontend/host-app/src/components/Skeleton/SkeletionTitle.js
