+++ frontend/host-app/src/libs/serviceWorker.js
@@ -83,8 +83,8 @@ function registerValidSW(swUrl, config) {
 						} else {
 							// At this point, everything has been precached.
 							// It's the perfect time to display a
-							// "Content is cached for offline use." message.
-							console.log("Content is cached for offline use.");
+							// "EventMonitor is cached for offline use." message.
+							console.log("EventMonitor is cached for offline use.");
 
 							// Execute callback
 							if (config && config.onSuccess) {
